refactor(T1c AC): impl-split SFAutoConnectService across 3 .cpp [BUILD-PASSED]
The AutoConnect slice, realized as a pure impl-split (one class, 3 .cpp) rather than the card's
extract-to-separate-services (deferred to optional architecture work) - same safe pattern as T8a.
SFAutoConnectService.cpp 4,771 -> 1,665; new _Belt.cpp 1,423 (ProcessSingleDistributor + belt
preview/build) + _Stackable.cpp 1,767 (stackable conveyor/pipe supports + power-pole). All <2k.
Zero behavior change (methods stay USFAutoConnectService members across TUs; maps/predicates intact;
no qualification/forwarders/header change). Shipping PackagePlugin clean (ExitCode=0). Solo lane;
auto-connect light-check folds into the next smoke.
